Revert "Merge pull request #69917 from CyrusNajmabadi/sgOOP2"
This reverts commit e1af8c0c91fd7257ded5711cd7e4366a1bb4ba83, reversing
changes made to 8f8abc373d058fc8abf1f66f677a4f9cd65bcf7e.
