+++ src/Workspaces/Core/Portable/Workspace/IWorkspaceConfigurationService.cs
@@ -44,11 +44,11 @@ namespace Microsoft.CodeAnalysis.Host
         [property: DataMember(Order = 3)] bool DisableRecoverableText = false,
         [property: DataMember(Order = 4)] bool ValidateCompilationTrackerStates =
 #if DEBUG // We will default this on in DEBUG builds
-            true,
+            true
 #else
-            false,
+            false
 #endif
-        [property: DataMember(Order = 5)] bool RunSourceGeneratorsInProcessOnly = false)
+        )
     {
         public WorkspaceConfigurationOptions()
             : this(CacheStorage: StorageDatabase.SQLite)
@@ -65,7 +65,6 @@ namespace Microsoft.CodeAnalysis.Host
             CacheStorage: StorageDatabase.None,
             EnableOpeningSourceGeneratedFiles: false,
             DisableSharedSyntaxTrees: false,
-            DisableRecoverableText: false,
-            RunSourceGeneratorsInProcessOnly: false);
+            DisableRecoverableText: false);
     }
 }
