+++ src/Workspaces/Remote/Core/ServiceDescriptors.cs
@@ -27,7 +27,6 @@ using Microsoft.CodeAnalysis.NavigateTo;
 using Microsoft.CodeAnalysis.NavigationBar;
 using Microsoft.CodeAnalysis.Rename;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
-using Microsoft.CodeAnalysis.SourceGeneration;
 using Microsoft.CodeAnalysis.StackTraceExplorer;
 using Microsoft.CodeAnalysis.SymbolSearch;
 using Microsoft.CodeAnalysis.TaskList;
@@ -84,7 +83,6 @@ namespace Microsoft.CodeAnalysis.Remote
             (typeof(IRemoteLegacySolutionEventsAggregationService), null),
             (typeof(IRemoteStackTraceExplorerService), null),
             (typeof(IRemoteUnitTestingSearchService), null),
-            (typeof(IRemoteSourceGenerationService), null),
         });
 
         internal readonly RemoteSerializationOptions Options;
