+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.cs.xlf
@@ -132,11 +132,6 @@
         <target state="translated">Poskytovatel prostředků</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">Průzkumník trasování zásobníku</target>
