+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.fr.xlf
@@ -132,11 +132,6 @@
         <target state="translated">Fournisseur de ressources</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">Explorateur de traces de pile</target>
