+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.it.xlf
@@ -132,11 +132,6 @@
         <target state="translated">Provider di asset</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">Esplora analisi dello stack</target>
