+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.ja.xlf
@@ -132,11 +132,6 @@
         <target state="translated">資産プロバイダー</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">スタック トレース エクスプローラー</target>
