+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.ko.xlf
@@ -132,11 +132,6 @@
         <target state="translated">자산 공급자</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">스택 추적 탐색기</target>
