+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.pt-BR.xlf
@@ -132,11 +132,6 @@
         <target state="translated">Provedor de ativos</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">Explorador de Rastreamento de Pilha</target>
