+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.ru.xlf
@@ -132,11 +132,6 @@
         <target state="translated">Поставщик ресурсов</target>
         <note />
       </trans-unit>
-      <trans-unit id="FeatureName_SourceGeneration">
-        <source>Source generation</source>
-        <target state="new">Source generation</target>
-        <note />
-      </trans-unit>
       <trans-unit id="FeatureName_StackTraceExplorer">
         <source>Stack Trace Explorer</source>
         <target state="translated">Обозреватель трассировки стека</target>
